Add files via upload
The code is a Python script that generates passwords and inserts them into an SQL Server database.
It uses the tkinter library to create a GUI and allows the user to select between two options for
password generation: 1) manual input of a password and its corresponding application and username,
or 2) automatic generation of a password with a specified length. The passwords are encrypted using
cryptography. fernet library before being inserted into the database. The program also verifies for
duplicates before inserting the password. also, the program encrypts the encryption keys with a master
key that is stored in the environment variables in the computer and stores them in an encrypted keys file